fix: types

+++ backend/repository/team.go
@@ -4,7 +4,7 @@ import "backend/types/database"
 
 type TeamRepository interface {
 	GetTeamById(uint64) (*database.Team, error)
-	ChangeTeamScore(uint64, float64) error
+	ChangeTeamScore(uint64, int32) error
 	GetTeams() ([]*database.Team, error)
 	GetPodium() ([]*database.Podium, error)
 	GetRanking() ([]*database.Ranking, error)
